주문 생성 케이스 작성

+++ src/main/java/com/quid/tdd/product/controller/model/ProductResponse.java
@@ -3,7 +3,7 @@ package com.quid.tdd.product.controller.model;
 import com.quid.tdd.product.domain.Product;
 import java.util.List;
 
-public record ProductResponse(Long id, String name, int price) {
+public record ProductResponse(Long id, String name, Long price) {
 
 
     public static ProductResponse of(Product product) {
