+++ src/main/java/com/quid/tdd/product/controller/model/UpdateProductRequest.java
@@ -1,5 +1,5 @@
 package com.quid.tdd.product.controller.model;
 
-public record UpdateProductRequest(Long id, String name, int price) {
+public record UpdateProductRequest(Long id, String name, Long price) {
 
 }
